Modernize SiStripMonitorClient
