+++ DQM/SiStripMonitorClient/plugins/SiStripDcsInfo.cc
@@ -1,20 +1,89 @@
+#include "DQM/SiStripCommon/interface/SiStripFolderOrganizer.h"
+#include "DQM/SiStripMonitorClient/interface/SiStripUtility.h"
+#include "DQMServices/Core/interface/DQMStore.h"
+#include "CalibFormats/SiStripObjects/interface/SiStripDetCabling.h"
+#include "CalibTracker/Records/interface/SiStripDetCablingRcd.h"
+#include "CondFormats/RunInfo/interface/RunInfo.h"
+#include "CondFormats/DataRecord/interface/RunSummaryRcd.h"
+#include "CondFormats/DataRecord/interface/SiStripCondDataRecords.h"
+#include "CondFormats/SiStripObjects/interface/SiStripDetVOff.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/ESWatcher.h"
+#include "FWCore/Framework/interface/LuminosityBlock.h"
+#include "FWCore/Framework/interface/Run.h"
+#include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-#include "DQMServices/Core/interface/DQMStore.h"
-#include "CondFormats/RunInfo/interface/RunInfo.h"
-
-#include "DQM/SiStripCommon/interface/SiStripFolderOrganizer.h"
-#include "SiStripDcsInfo.h"
-#include "DQM/SiStripMonitorClient/interface/SiStripUtility.h"
+#include "Geometry/Records/interface/TrackerTopologyRcd.h"
 
 #include <iostream>
 #include <iomanip>
 #include <cstdio>
-#include <string>
 #include <sstream>
 #include <cmath>
+#include <fstream>
+#include <string>
+#include <vector>
+#include <map>
+#include <unordered_map>
+
+class SiStripDetVOff;
+class SiStripDetCabling;
+class RunInfo;
+
+class SiStripDcsInfo : public edm::one::EDAnalyzer<> {
+public:
+  typedef dqm::harvesting::MonitorElement MonitorElement;
+  typedef dqm::harvesting::DQMStore DQMStore;
+
+  SiStripDcsInfo(const edm::ParameterSet& ps);
+
+private:
+  void beginJob() override;
+  void beginRun(edm::Run const& run, edm::EventSetup const& eSetup);
+  void beginLuminosityBlock(edm::LuminosityBlock const& lumiSeg, edm::EventSetup const& eSetup);
+  void endLuminosityBlock(edm::LuminosityBlock const& lumiSeg, edm::EventSetup const& iSetup);
+  void endRun(edm::Run const& run, edm::EventSetup const& eSetup);
+  void analyze(edm::Event const&, edm::EventSetup const&) override;
+
+  void bookStatus(DQMStore& dqm_store);
+  void readStatus(edm::EventSetup const&, int transition);
+  void readCabling(edm::EventSetup const&);
+  void addBadModules(DQMStore& dqm_store);
+  void fillStatus(DQMStore& dqm_store);
+  void fillDummyStatus(DQMStore& dqm_store);
+
+  MonitorElement* DcsFraction_{nullptr};
+
+  struct SubDetMEs {
+    std::string folder_name;
+    MonitorElement* DcsFractionME;
+    int TotalDetectors;
+    std::vector<uint32_t> FaultyDetectors;
+    std::unordered_map<uint32_t, uint16_t> NLumiDetectorIsFaulty;
+  };
+
+  std::map<std::string, SubDetMEs> SubDetMEsMap{};
+  unsigned long long m_cacheIDDcs_{};
+  bool bookedStatus_{false};
+
+  int nFEDConnected_{};
+
+  int nLumiAnalysed_{};
+
+  bool IsLumiGoodDcs_{false};
+  int nGoodDcsLumi_{};
+  static constexpr float MinAcceptableDcsDetFrac_{0.90};
+  static constexpr float MaxAcceptableBadDcsLumi_{2};
+
+  const SiStripDetCabling* detCabling_;
+  edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> tTopoToken0_, tTopoToken1_, tTopoToken2_, tTopoToken3_;
+  edm::ESGetToken<SiStripDetVOff, SiStripDetVOffRcd> detVOffToken0_, detVOffToken1_, detVOffToken2_;
+  edm::ESWatcher<SiStripFedCablingRcd> fedCablingWatcher_;
+  edm::ESGetToken<SiStripDetCabling, SiStripDetCablingRcd> detCablingToken_;  // beginRun
+  edm::ESGetToken<RunInfo, RunInfoRcd> runInfoToken_;
+};
 
 //
 // -- Contructor
